pyproject.toml
[tool.poetry]
name = "lms-project"
version = "0.1.0"
description = ""
authors = ["TagRomEdu <qwertystyles@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
ipython = "^8.17.2"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
coverage = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
